+++ rust/sp-rs/Cargo.toml
@@ -1,0 +1,26 @@
+[package]
+name = "sp-rs"
+version = "0.1.0"
+authors = ["xthexder", "nauscar"]
+edition = "2021"
+license-file = "../../LICENSE-MPL-2.0"
+
+[features]
+default = ["gfx", "wasm"]
+api = []
+gfx = ["dep:vulkano", "dep:vulkano-win", "dep:winit"]
+wasm = ["dep:wasmer", "dep:static_assertions"]
+
+[dependencies]
+cxx = "1.0"
+static_assertions = { version = "1.1", optional = true }
+vulkano = { version = "0.33", optional = true }
+vulkano-win = { version = "0.33", optional = true }
+winit = { version = "0.28", optional = true }
+wasmer = { version = "2.1", optional = true }
+
+[build-dependencies]
+cxx-build = "1.0"
+
+[lib]
+crate-type = ["staticlib"]
