+++ rust/sp-rs/include/api.hh
@@ -1,0 +1,13 @@
+#pragma once
+#include "rust/cxx.h"
+
+namespace sp {
+    namespace api {
+        class {
+            Api {
+            public:
+                Api();
+            }
+        }
+    } // namespace api
+} // namespace sp
